refactor(#8/user): dto 패키지 위치 이동
