+++ src/main/java/org/sopt/post/presentation/response/GetDetailedPostResponse.java
@@ -1,7 +1,7 @@
 package org.sopt.post.presentation.response;
 
 import org.sopt.post.domain.Post;
-import org.sopt.user.dto.response.GetUserResponse;
+import org.sopt.user.presentation.dto.response.GetUserResponse;
 
 public record GetDetailedPostResponse(Long id, GetUserResponse user, String title, String content) {
     public static GetDetailedPostResponse from(Post post) {
