+++ src/main/java/org/sopt/post/presentation/response/GetSimplePostResponse.java
@@ -1,7 +1,7 @@
 package org.sopt.post.presentation.response;
 
 import org.sopt.post.domain.Post;
-import org.sopt.user.dto.response.GetUserResponse;
+import org.sopt.user.presentation.dto.response.GetUserResponse;
 
 public record GetSimplePostResponse(Long id, GetUserResponse author, String title) {
     public static GetSimplePostResponse from(Post post) {
